+++ trunk/ndk/strace.cpp
@@ -2,7 +2,7 @@
 
 namespace ndk
 {
-size_t strace::count_ = 0;
+unsigned int strace::count_ = 0;
 thread_mutex strace::mutex_;
 } // namespace ndk
